+++ stripe-core/tests/Web/Stripe/Test/ApplicationFee.hs
@@ -1,9 +1,9 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RebindableSyntax #-}
-module Test.ApplicationFee where
+module Web.Stripe.Test.ApplicationFee where
 
 import Test.Hspec
-import Test.Prelude
+import Web.Stripe.Test.Prelude
 import Web.Stripe.ApplicationFee
 import Data.Either
