+++ stripe-core/tests/Web/Stripe/Test/Recipient.hs
@@ -1,11 +1,11 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE RebindableSyntax #-}
-module Test.Recipient where
+module Web.Stripe.Test.Recipient where
 
 import           Data.Either
 import           Test.Hspec
-import           Test.Prelude
+import           Web.Stripe.Test.Prelude
 import           Web.Stripe.Recipient
 
 recipientTests :: StripeSpec
